Refactoring Runtime.Api - InvalidatedExpressions
Moving out of the `Runtime` class.
